requirements.txt
flask
flask-socketio
gevent
gym
matplotlib
numpy
pillow
torch
